requirements.txt
biopython==1.78
JPype1==1.2.1
numpy==1.20.1
osprey==3.2
click==8.0.1
pyyaml